requirements.txt
qrcode[pil]
Pillow
pyyaml
